refactor(sdk): extract @plot/sdk and make it own shared work shapes
The published plot-ai/sdk contract gets its own package. @plot/agent imports
WorkDisplay/OperatorAction/OperatorActionConfirm type-only from @plot/sdk instead
of re-declaring them; the extension loader's plot-ai/sdk virtual module and the
release sdk build now point at packages/sdk.
