+++ packages/session/package.json
@@ -52,10 +52,6 @@
 			"types": "./src/protocol-codec.ts",
 			"import": "./src/protocol-codec.ts"
 		},
-		"./sdk": {
-			"types": "./src/sdk.ts",
-			"import": "./src/sdk.ts"
-		},
 		"./workflow": {
 			"types": "./src/workflow.ts",
 			"import": "./src/workflow.ts"
@@ -64,14 +60,14 @@
 	"scripts": {
 		"typecheck": "tsc -p tsconfig.json --noEmit",
 		"lint": "oxlint . --tsconfig tsconfig.json",
-		"test": "bun test test",
-		"build:sdk": "rm -rf dist && bun build src/sdk.ts --target=node --format=esm --outfile=dist/sdk.js && tsc -p tsconfig.sdk.json"
+		"test": "bun test test"
 	},
 	"dependencies": {
 		"@earendil-works/pi-ai": "0.79.0",
 		"@earendil-works/pi-coding-agent": "0.79.0",
 		"@plot/agent": "workspace:*",
 		"@plot/common": "workspace:*",
+		"@plot/sdk": "workspace:*",
 		"effect": "4.0.0-beta.78",
 		"eta": "4.6.0",
 		"jiti": "^2.7.0",
