+++ scripts/release/build.ts
@@ -17,7 +17,7 @@ import {
 	npmPackageDir,
 	packageTemplate,
 	repoDir,
-	sessionPackageDir,
+	sdkPackageDir,
 	releaseDir,
 	releaseTargets,
 	version,
@@ -28,7 +28,7 @@ mkdirSync(releaseDir, { recursive: true });
 
 await $`bun --filter @plot/web build`.cwd(repoDir);
 await $`bun run scripts/web-assets.ts`.cwd(repoDir);
-await $`bun run build:sdk`.cwd(sessionPackageDir);
+await $`bun run build:sdk`.cwd(sdkPackageDir);
 await Promise.all(releaseTargets.map((target) => buildPlatformPackage(target)));
 await buildUmbrellaPackage();
 
@@ -85,11 +85,11 @@ async function buildUmbrellaPackage() {
 		recursive: true,
 	});
 	cpSync(
-		join(sessionPackageDir, "dist", "sdk.js"),
+		join(sdkPackageDir, "dist", "sdk.js"),
 		join(packageDir, "lib", "sdk.js"),
 	);
 	cpSync(
-		join(sessionPackageDir, "dist", "sdk.d.ts"),
+		join(sdkPackageDir, "dist", "sdk.d.ts"),
 		join(packageDir, "lib", "sdk.d.ts"),
 	);
 	cpSync(
